# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadResponse {
  /** Unique session ID for this upload */
  sessionId: string;
  /** Original filename */
  filename: string;
  /** Video duration in seconds */
  duration: number;
  /** Video width in pixels */
  width?: number;
  /** Video height in pixels */
  height?: number;
  /** File size in bytes */
  size: number;
}

/**
 * Extraction mode
 */
export type ExtractRequestMode =
  (typeof ExtractRequestMode)[keyof typeof ExtractRequestMode];

export const ExtractRequestMode = {
  timestamp: "timestamp",
  interval: "interval",
} as const;

export interface ExtractRequest {
  /** Session ID from upload */
  sessionId: string;
  /** Extraction mode */
  mode: ExtractRequestMode;
  /** Specific timestamp (hh:mm:ss) for timestamp mode */
  timestamp?: string;
  /** Interval in seconds for interval mode */
  interval?: number;
  /**
   * JPEG quality 1-100
   * @minimum 1
   * @maximum 100
   */
  quality?: number;
}

export interface Frame {
  id: string;
  filename: string;
  url: string;
  /** Timestamp in seconds */
  timestamp: number;
  /** Whether this frame is suggested (sharp and bright) */
  suggested: boolean;
  /** Sharpness score 0-100 */
  sharpness: number;
  /** Brightness score 0-100 */
  brightness: number;
}

export interface ExtractResponse {
  sessionId: string;
  frameCount: number;
  frames: Frame[];
}

export interface FramesResponse {
  sessionId: string;
  frames: Frame[];
}

export interface DownloadZipRequest {
  sessionId: string;
  /** List of frame IDs to include in ZIP */
  frameIds: string[];
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type UploadVideoBody = {
  video: Blob;
};

export type GetFramesParams = {
  /**
   * Session ID from upload response
   */
  sessionId: string;
};