# lib/api-zod/src/generated/types/downloadZipRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DownloadZipRequest {
  sessionId: string;
  /** List of frame IDs to include in ZIP */
  frameIds: string[];
}