# lib/api-zod/src/generated/types/extractRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractRequestMode } from "./extractRequestMode";

export interface ExtractRequest {
  /** Session ID from upload */
  sessionId: string;
  /** Extraction mode */
  mode: ExtractRequestMode;
  /** Specific timestamp (hh:mm:ss) for timestamp mode */
  timestamp?: string;
  /** Interval in seconds for interval mode */
  interval?: number;
  /**
   * JPEG quality 1-100
   * @minimum 1
   * @maximum 100
   */
  quality?: number;
}