# lib/api-zod/src/generated/types/extractResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Frame } from "./frame";

export interface ExtractResponse {
  sessionId: string;
  frameCount: number;
  frames: Frame[];
}