# lib/api-zod/src/generated/types/frame.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Frame {
  id: string;
  filename: string;
  url: string;
  /** Timestamp in seconds */
  timestamp: number;
  /** Whether this frame is suggested (sharp and bright) */
  suggested: boolean;
  /** Sharpness score 0-100 */
  sharpness: number;
  /** Brightness score 0-100 */
  brightness: number;
}