# lib/api-zod/src/generated/types/framesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Frame } from "./frame";

export interface FramesResponse {
  sessionId: string;
  frames: Frame[];
}