# lib/api-zod/src/generated/types/uploadResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FrameSnap API specification
 * OpenAPI spec version: 0.1.0
 */

export interface UploadResponse {
  /** Unique session ID for this upload */
  sessionId: string;
  /** Original filename */
  filename: string;
  /** Video duration in seconds */
  duration: number;
  /** Video width in pixels */
  width?: number;
  /** Video height in pixels */
  height?: number;
  /** File size in bytes */
  size: number;
}